Refactor from Duke to Tommy
